Simplified testing moved to test/fixtures + serverspec testing
Updated default recipe to remove use of node.name
ISSUE: bug seen on a second converge. Where it would set the hostname to test.test.example.com
This replaces the use of node.name to node['hostname'] as node.name is not guarenteed to contain the node FQDN nor do we want the FQDN here we hostname the hostname to avoid the idempotent bug. Replaced use of only_if to not_if as the guard was not workingthe use of a lazy block is superfluous as ohai will not reload the hostname even after an explicit reload.